add a showPlayers page (equivalent functionality to showTeams).

+++ src/main/java/com/ilmservice/fantasyfootball/db/repositories/PlayerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.ilmservice.fantasyfootball.db.entities.Player;
 
-public interface PlayerRepository extends CrudRepository<Player, Long> {
+public interface PlayerRepository extends CrudRepository<Player, Integer> {
 
 }
